requirements.txt
python-dotenv
matplotlib
numpy
pandas
requests
scikit-learn
seaborn
jupyter
xgboost
plotly
kaleido
holidays
pyarrow